+++ tor-0.2.4.21/src/or/relay.h
@@ -13,9 +13,9 @@
 #define TOR_RELAY_H
 
 extern uint64_t stats_n_relay_cells_relayed;
-extern uint64_t stats_n_relay_cells_delivered;
+extern int stats_n_relay_cells_delivered;
 extern int stats_n_sendme_cell;
-uint64_t get_stats_sendme();
+int get_stats_sendme();
 int circuit_receive_relay_cell(cell_t *cell, circuit_t *circ,
                                cell_direction_t cell_direction);
